+++ tests/linearGaussian_snpe_test.py
@@ -288,7 +288,6 @@ def test_c2st_multi_round_snpe_on_linearGaussian(method_str: str, set_seed):
 
 # Testing rejection and mcmc sampling methods.
 @pytest.mark.slow
-@pytest.mark.parametrize("snpe_method", [SNPE_A, SNPE_C])
 @pytest.mark.parametrize(
     "sample_with, mcmc_method, prior_str",
     (
@@ -299,9 +298,7 @@ def test_c2st_multi_round_snpe_on_linearGaussian(method_str: str, set_seed):
         ("rejection", "rejection", "uniform"),
     ),
 )
-def test_api_snpe_c_posterior_correction(
-    snpe_method: type, sample_with, mcmc_method, prior_str, set_seed
-):
+def test_api_snpe_c_posterior_correction(sample_with, mcmc_method, prior_str, set_seed):
     """Test that leakage correction applied to sampling works, with both MCMC and
     rejection.
 
@@ -326,7 +323,7 @@ def test_api_snpe_c_posterior_correction(
     simulator, prior = prepare_for_sbi(
         lambda theta: linear_gaussian(theta, likelihood_shift, likelihood_cov), prior
     )
-    inference = snpe_method(
+    inference = SNPE_C(
         prior,
         simulation_batch_size=50,
         sample_with=sample_with,
@@ -432,7 +429,7 @@ def test_sample_conditional(snpe_method: type, set_seed):
     error = np.abs(sample_kde_grid - eval_grid.numpy())
 
     max_err = np.max(error)
-    assert max_err < 0.0025
+    assert max_err < 0.0026
 
 
 @pytest.mark.parametrize("snpe_method", [SNPE_A, SNPE_C])